Update to support newer charmcraft and pin old ops.
For now, I still want to have a Bionic version of the charm, so I want to keep using ops 1.5.
This seems to do what I want, with trimmed down to no longer use a subproject and just specify
the dependencies in requirements.txt. Charmcraft 2.3.0 successfully builds this project and 2.9
seems to deploy it fine. Still need to work out how 3.* will handle having a metadata series and
a manifest.yaml