fix: update to ruff and ops 2.17 and ops-scenario
Lots of small changes, nothing very fundamental.
Update to new ops and build infrastructure, which gets us to newer ruff.
Wants a pyproject.toml. For now, I'm avoiding tox, but it does mean that
I have to leave in the crummy env manipulation to teach the test suite
how to find the source directory. (I could put it as environment
variables in the make file, I was hoping for something standard in
pytest, but I haven't found it yet.)